+++ packages/expo-dev-menu/ios/DevMenuAppInstance.swift
@@ -1,9 +1,6 @@
 // Copyright 2015-present 650 Industries. All rights reserved.
 
 import React
-#if canImport(ReactAppDependencyProvider)
-import ReactAppDependencyProvider
-#endif
 
 @objc
 class DevMenuAppInstance: DevMenuRCTAppDelegate {
@@ -17,7 +14,7 @@ class DevMenuAppInstance: DevMenuRCTAppDelegate {
 
     super.init()
 #if canImport(ReactAppDependencyProvider)
-	self.dependencyProvider = RCTAppDependencyProvider()
+    self.dependencyProvider = EXAppDependencyProvider()
 #endif
     super.initRootViewFactory()
   }